refact: split chat service in 2 parts
