+++ backend/internal/constants/message.go
@@ -1,0 +1,9 @@
+package constants
+
+type MessageType string
+
+const (
+	MessageTypeUser      MessageType = "user"
+	MessageTypeAssistant MessageType = "assistant"
+	MessageTypeSystem    MessageType = "system"
+)
